+++ CHANGELOG.md
@@ -122,6 +122,8 @@
   - Update stable-fast with support for torch 2.2.2 and 2.3.0, thanks @Aptronymist
   - Add torch *cudaMallocAsync* in compute options  
     Can improve memory utilization on compatible GPUs (RTX and newer)  
+  - Torch dynamic profiling  
+    You can enable/disable full torch profiling in settings top menu on-the-fly  
   - Support controlnet manually downloads models in both standalone and diffusers format  
     For standalone, simply copy safetensors file to `models/control/controlnet` folder  
     For diffusers format, create folder with model name in `models/control/controlnet/`